+++ library/Makefile
@@ -159,6 +159,7 @@ OBJS_X509= \
 	   x509_crl.o \
 	   x509_crt.o \
 	   x509_csr.o \
+	   x509_info.o \
 	   x509write_crt.o \
 	   x509write_csr.o \
 	   # This line is intentionally left blank
